requirements.txt
asgiref==3.2.5
certifi==2019.11.28
Django==3.0.4
django-filter==2.2.0
flake8==3.8.3
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.1
